Support replicateH on HNil
This changes the `ReplicateH` type class to line up a little better with the
`Sequencer` type class.

This also adds implicit `Sequencer` instances for HNil. One can sequence an
empty `List`, so I can't think of any reason that they shouldn't be able to
sequence an `HNil`. Adding this support allows one to pass zero to `replicateH`.