Fix issues with Empty derivation for coproducts
When deriving an `Empty` instance for coproducts we look for a
unique member of the coproduct with an `Empty` instance.

The way to ensure this is collect all available instances for
members of the coproduct in an `HList` and assert the result has
exactly one element. For this purpose we add the `LiftSome` typeclass.